// embedded-components/src/api/generated/ep-transactions-v3.schemas.ts
export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  PENDING: 'PENDING',
  CANCELED: 'CANCELED',
  COMPLETED: 'COMPLETED',
  COMPLETED_WITH_EXCEPTIONS: 'COMPLETED_WITH_EXCEPTIONS',
  COMPLETED_NOT_IN_TARGET_WALLET: 'COMPLETED_NOT_IN_TARGET_WALLET',
  REJECTED: 'REJECTED',
  RETURNED: 'RETURNED',
  PARTIALLY_COMPLETED: 'PARTIALLY_COMPLETED',
  UNDEFINED: 'UNDEFINED',
} as const;

/**
 * The transaction type indicates the nature of the transaction. Certain types cannot be initiated through the API and, as a result, are excluded from the transaction request type field. API and, as a result, are excluded from the transaction request type field.
 */
export type PaymentTypeResponseV3 =
  (typeof PaymentTypeResponseV3)[keyof typeof PaymentTypeResponseV3];

export const PaymentTypeResponseV3 = {
  RTP: 'RTP',
  WIRE: 'WIRE',
  ACH: 'ACH',
  TRANSFER: 'TRANSFER',
  REVERSAL: 'REVERSAL',
  RETURN: 'RETURN',
  OTHER: 'OTHER',
  CARD: 'CARD',
  FEE: 'FEE',
  MISCELLANEOUS_CREDIT: 'MISCELLANEOUS_CREDIT',
  MISCELLANEOUS_DEBIT: 'MISCELLANEOUS_DEBIT',
} as const;

/**
 * Details of the originating transaction for fees, returns, or linked transactions.
 */
export interface OriginatingTransactionV3 {
  /** Unique identifier of the originating transaction. */
  id?: string;
  type?: PaymentTypeResponseV3;
}

/**
 * Underlying reason for the payment transaction
 */
export interface TransactionPaymentPurposeV3 {
  /**
   * Purpose of the payment in coded form. Must be between 1 and 4 characters.
   * @maxLength 4
   */
  code?: string;
  /**
   * Purpose of the payment in proprietary form. Must be between 1 and 35 characters.
   * @maxLength 35
   */
  customCode?: string;
}

/**
 * Foreign exchange information for cross-border transactions.
 */
export interface TransactionFxInformation {
  /**
   * Unique identifier for the foreign exchange rate
   * @minLength 1
   */
  rateId?: string;
  /**
   * The rate at which you charge for an FX transfer. You must express this as a decimal figure. For example, if you take a commission of 10%, the spread is expressed as 0.1
   * @minLength 1
   */
  clientSpread?: string;
}

/**
 * Foreign exchange information for cross-border transactions.
 */
export type GetTransactionFxInformationV3 = TransactionFxInformation & {
  /** Unique identifier for the foreign exchange contract */
  contractId?: string;
  /** The exchange rate applied to the transaction */
  exchangeRate?: string;
};
// ...
export interface TransactionCounterPartyDetailsV3 {
  /** The type of the counterparty. */
  type?: string;
  /** The name of the counterparty (individual or organization). */
  name?: string;
  address?: TransactionCounterPartyAddressDetailsV3;
  /**
   * The Identity details for the counterparty.
   * @minItems 0
   */
  identities?: TransactionCounterPartyIdentityDetailsV3[];
  contacts?: TransactionCounterPartyContactDetailsV3;
}
// ...
export type TransactionCounterPartyAccountV3 =
  | TransactionCounterPartyExternalAccountV3
  | TransactionCounterPartyRegisteredAccountV3
  | TransactionCounterPartyCardAccountV3;

export interface GetTransactionCounterPartyV3 {
  /** The id of the counterparty. */
  id?: string;
  /** The client identifier associated with the counterparty. */
  clientId?: string;
  /** The type of the counterparty. */
  type?: string;
  partyDetails?: TransactionCounterPartyDetailsV3;
  account?: TransactionCounterPartyAccountV3;
}
// ...
export interface ApiErrorReasonV2 {
  /**
   * Short code that identifies the error - publicly cataloged and documented
   * @minLength 0
   * @maxLength 5
   * @nullable
   */
  code?: string | null;
  /**
   * Part of the request which is responsible for the reason
   * @nullable
   */
  location?: ApiErrorReasonV2Location;
  /**
   * The location of the property or parameter in error
   * @minLength 0
   * @maxLength 512
   * @nullable
   */
  field?: string | null;
  /**
   * Message describing the reason. This message can typically be displayed to your platform's users, except in cases specified otherwise
   * @minLength 0
   * @maxLength 2048
   */
  message: string;
}

/**
 * Schema representing a version 2 API error, including title, HTTP status, and additional context.
 */
export interface ApiErrorV2 {
  /**
   * Short humanly-readable title of the error
   * @minLength 1
   * @maxLength 512
   */
  title: string;
  /**
   * HTTP status code
   * @minimum 100
   * @maximum 599
   */
  httpStatus?: number;
  /**
   * Internal assigned traced identifier
   * @minLength 0
   * @maxLength 512
   */
  traceId?: string;
  /**
   * Client provided request identifier
   * @minLength 0
   * @maxLength 512
   */
  requestId?: string;
  /**
   * Provides additional context and detail on the validation errors
   * @minItems 0
   * @maxItems 100
   */
  context?: ApiErrorReasonV2[];
}

export interface TransactionGetResponseV3 {
  /** Unique identifier for the transaction. */
  id: string;
  /** Transaction reference id provided by client. */
  transactionReferenceId?: string;
  status: TransactionStatus;
  type: PaymentTypeResponseV3;
  /**
   * The date the client requested for execution.
   * @minLength 1
   */
  requestedExecutionDate?: string;
  /** The date funds are considered available to the creditor. */
  valueDate?: string;
  /** The code that was used when creating the transaction. */
  localInstrumentCode?: string;
  /** Transaction information that states its purpose for easier tracking. */
  memo?: string;
  /** The value of the transaction. */
  amount?: string;
  /** The currency of the transaction in ISO 4217 Alpha-3 Currency Code format. */
  currency: string;
  /**
   * Transaction value in the target currency credited to the creditor.
   * @maxLength 40
   * @pattern ^\d+(\.\d+)?$
   */
  targetAmount?: string;
  /**
   * Currency of the creditor account.
   * @maxLength 3
   * @pattern ^[A-Z]{3}$
   */
  targetCurrency?: string;
  originatingTransaction?: OriginatingTransactionV3;
  paymentPurpose?: TransactionPaymentPurposeV3;
  fxInformation?: GetTransactionFxInformationV3;
  debtor?: GetTransactionCounterPartyV3;
  creditor?: GetTransactionCounterPartyV3;
  error?: ApiErrorV2;
}
// ...
export interface TransactionResponseV3 {
  /**
   * The resource Id of your transaction to use when looking up details of this transaction.
   * @minLength 1
   */
  id?: string;
  /**
   * Transaction reference ID provided by client
   * @minLength 1
   */
  transactionReferenceId: string;
  /**
   * ISO date format - yyyy-MM-dd
   * @minLength 1
   */
  requestedExecutionDate?: string;
}